Relocated route components to `pages` folder and updated components to use v6 conventions
