+++ src/pages/blog.js
@@ -1,0 +1,50 @@
+import React from 'react'
+import { SiteData, Head, RouteData } from 'react-static'
+import { Link } from '@reach/router'
+//
+import { withStyles } from '@material-ui/core/styles'
+import Grid from '@material-ui/core/Grid'
+import Paper from '@material-ui/core/Paper'
+import Typography from '@material-ui/core/Typography'
+import Moment from 'react-moment'
+import ReactMarkdown from 'react-markdown'
+
+const styles = theme => ({
+  root: {
+  },
+  post: {
+    background: '#5a5a5a',
+    padding: theme.spacing.unit * 2,
+  },
+  link: {
+    textDecoration: 'none',
+  },
+})
+
+export default withStyles(styles)(({ classes }) => (
+  <div className={classes.root}>
+    <SiteData>
+      {({title}) => (
+        <Head title={`Blog - ${title}`} />
+      )}
+    </SiteData>
+    <RouteData>
+      {({header, posts}) => (
+        <React.Fragment>
+          <Typography variant="h4" paragraph>{header || 'All Posts'}</Typography>
+          <Grid container direction="column" spacing={16}>
+            {posts.map(post => (
+              <Grid item xs="auto" key={post.path}>
+                <Paper className={classes.post}>
+                  <Typography variant="h5" component={Link} to={post.path} className={classes.link}>{post.title}</Typography>
+                  <Typography variant="subtitle1" paragraph><Moment date={post.date} format="dddd, MMMM Do, YYYY" /></Typography>
+                  <Typography variant="body2" color="textSecondary" component={ReactMarkdown} source={post.body} paragraph />
+                </Paper>
+              </Grid>
+            ))}
+          </Grid>
+        </React.Fragment>
+      )}
+    </RouteData>
+  </div>
+))
