+++ src/pages/index.js
@@ -1,5 +1,6 @@
 import React from 'react'
-import { SiteData, Head, RouteData, Link } from 'react-static'
+import { SiteData, Head, RouteData } from 'react-static'
+import { Link } from '@reach/router'
 //
 import { withStyles } from '@material-ui/core/styles'
 import Typography from '@material-ui/core/Typography'
@@ -18,9 +19,11 @@ const styles = theme => ({
 
 export default withStyles(styles)(({ classes }) => (
   <div className={classes.root}>
-    <SiteData render={({title}) => (
+    <SiteData>
+      {({title}) => (
         <Head title={title} />
-    )} />
+      )}
+    </SiteData>
     <Typography variant="h4" paragraph>Welcome!</Typography>
     <Typography variant="body2" paragraph>This site has nothing on it... yet.</Typography>
     <Typography variant="body1">Check out the <Link to="/blog">blog</Link>.</Typography>
